Separated inspector from RovClient
